Code cleanup. Removed test files from directory. Renamed h2_platform to platform to be more general. Reduced yaml file to the smallest it can be to run this orbit model. Annotated units everywhere relevant
