--- hopp/offshore/example_fixed_project_h2.yaml
@@ -1,21 +1,0 @@
-# Modified orbit configuration file for custom h2 platform
-#commissioning: 0.01
-#decommissioning: 0.15
-design_phases:
-- FixedPlatformDesign
-install_phases:
-  FixedPlatformInstallation: 0
-oss_install_vessel: example_heavy_lift_vessel
-plant:
-  capacity: 400
-site:
-  depth: 22.5
-  distance: 124
-  distance_to_landfall: 35
-  mean_windspeed: 9
-h2_platform:
-  tech_required_area: 300. # m**2
-  tech_combined_mass: 1000 # tonnes
-project_parameters:
-  project_lifetime: 30 # years 
-  opex_rate: 125       # $/kW/year
